Common steps introduced. Minor fixes.
