--- src/main/java/org/othertests/example/QueryLimit.java
@@ -1,14 +1,0 @@
-package org.othertests.example;
-
-public class QueryLimit {
-
-    private int limit;
-
-    public int getLimit() {
-        return limit;
-    }
-
-    public void setLimit(int limit) {
-        this.limit = limit;
-    }
-}
